backend/requirements.txt
fastapi[all]
uvicorn[standard]
beanie

python-decouple

PyJWT
passlib[bcrypt]

Jinja2
email-validator
fastapi-mail

fastapi_sso

pyotp

openai
elevenlabs
google-api-python-client